Fixing MSVC build.

+++ include/lightsky/draw/Animation.h
@@ -16,20 +16,25 @@
 /*-----------------------------------------------------------------------------
  * Forward declaration of math types not instantiated in the header.
 -----------------------------------------------------------------------------*/
-namespace ls {
-namespace math {
-    template <typename num_t> struct vec3_t;
-    template <typename num_t> struct quat_t;
+namespace ls
+{
+namespace math
+{
+template <typename num_t> struct vec3_t;
+template <typename num_t> struct quat_t;
 } // end math namespace
 } // end ls namespace
 
-namespace ls {
-namespace draw {
+namespace ls
+{
+namespace draw
+{
 
 /*-----------------------------------------------------------------------------
  * Forward declaration of scene types
 -----------------------------------------------------------------------------*/
 struct SceneNode;
+
 class SceneGraph;
 
 
@@ -41,7 +46,8 @@ class SceneGraph;
  * times an Animation should play. It is used by animationPlayer objects to
  * determine if an Animation plays once or multiple times.
 -------------------------------------*/
-enum class animation_play_t : unsigned {
+enum class animation_play_t : unsigned
+{
     ANIM_PLAY_ONCE,
     ANIM_PLAY_REPEAT,
 
@@ -56,11 +62,12 @@ enum class animation_play_t : unsigned {
  * This class keeps track of a single Animation, made up of 'tracks' or
  * keyframes, that are used to animate one or more meshes.
 ------------------------------------------------------------------------------*/
-class Animation {
-    
-  // Allow scene graphs to delete animation tracks based on the scene node
-  friend class SceneGraph;
-    
+class Animation
+{
+
+    // Allow scene graphs to delete animation tracks based on the scene node
+    friend class SceneGraph;
+
   private:
     /**
      * @brief playMode is by Animation players to determine if an Animation
@@ -96,20 +103,20 @@ class Animation {
      * @brief animationIds contains ID of the std::vector<AnimationChannel>
      * which will be used to identify a SceneNode's animation channel to use.
      */
-    std::vector<uint32_t> animationIds;
+    std::vector<size_t> animationIds;
 
     /**
      * @brief nodeTrackIds are used after the animationIds to determine the
      * exact AnimationChannel in a list of animation channels to use for an
      * animation.
      */
-    std::vector<uint32_t> nodeTrackIds;
-    
+    std::vector<size_t> nodeTrackIds;
+
     /**
      * @brief transformIds contains the indices of all node transformations
      * that will contain the resulting transformation after an animation.
      */
-    std::vector<uint32_t> transformIds;
+    std::vector<size_t> transformIds;
 
   public: // public member functions
     /**
@@ -158,7 +165,7 @@ class Animation {
      *
      * @return A reference to *this.
      */
-    Animation& operator =(const Animation& a) noexcept;
+    Animation& operator=(const Animation& a) noexcept;
 
     /**
      * @brief Move Operator
@@ -172,7 +179,7 @@ class Animation {
      *
      * @return A reference to *this.
      */
-    Animation& operator =(Animation&& a) noexcept;
+    Animation& operator=(Animation&& a) noexcept;
 
     /**
      * @brief Determine the current play mode used by *this.
@@ -198,7 +205,7 @@ class Animation {
      * @return An unsigned integer, containing the hashed value of *this
      * Animation's name.
      */
-    unsigned get_anim_id() const noexcept;
+    size_t get_anim_id() const noexcept;
 
     /**
      * @brief Retrieve the name of *this Animation.
@@ -268,7 +275,7 @@ class Animation {
      * @return A reference to a constant vector of indices which reference
      * the "currentTransform" objects in a SceneGraph.
      */
-    const std::vector<uint32_t>& get_transforms() const noexcept;
+    const std::vector<size_t>& get_transforms() const noexcept;
 
     /**
      * @brief Retrieve the list of indices which will be used to reference a
@@ -284,7 +291,7 @@ class Animation {
      * per-node animation channels
      * (SceneGraph::nodeAnims[animTrackId][nodeTrackId]).
      */
-    const std::vector<uint32_t>& get_node_tracks() const noexcept;
+    const std::vector<size_t>& get_node_tracks() const noexcept;
 
     /**
      * @brief Retrieve the list of indices which will be used to reference
@@ -297,7 +304,7 @@ class Animation {
      * reference an array of AnimationChannel objects in a SceneGraph.
      * (SceneGraph::nodeAnims[animTrackId]).
      */
-    const std::vector<uint32_t>& get_node_animations() const noexcept;
+    const std::vector<size_t>& get_node_animations() const noexcept;
 
     /**
      * @brief Get the number of Animation channels that will be animated by
@@ -306,8 +313,8 @@ class Animation {
      * @return The total number of node channels which *this Animation object
      * runs during any given frame.
      */
-    unsigned get_num_anim_channels() const noexcept;
-    
+    size_t get_num_anim_channels() const noexcept;
+
     /**
      * @brief Add an Animation channel to *this.
      * 
@@ -318,7 +325,7 @@ class Animation {
      * An unsigned integer, containing the index of the AnimationChannel in
      * the input node's std::vector<AnimationChannel> to use for animation.
      */
-    void add_anim_channel(const SceneNode& node, const uint32_t nodeTrackId) noexcept;
+    void add_anim_channel(const SceneNode& node, const size_t nodeTrackId) noexcept;
 
     /**
      * Remove a single Animation channel from *this.
@@ -326,18 +333,18 @@ class Animation {
      * @param trackId
      * The index of the Animation channel to remove.
      */
-    void remove_anim_channel(const unsigned trackId) noexcept;
+    void remove_anim_channel(const size_t trackId) noexcept;
 
     /**
      * Remove all Animation keyframes and channels inside of *this.
      */
     void clear_anim_channels() noexcept;
-    
+
     /**
      * @brief Reserve a number of animation channels to help avoid the chances
      * of a reallocation when adding single animations.
      */
-    void reserve_anim_channels(const unsigned reserveSize) noexcept;
+    void reserve_anim_channels(const size_t reserveSize) noexcept;
 
     /**
      * @brief Animate nodes in a sceneGraph.
@@ -354,7 +361,7 @@ class Animation {
      * assertion will be raised if this value is less than 0.0.
      */
     void animate(SceneGraph& graph, const anim_prec_t percentDone) const noexcept;
-    
+
     /**
      * Initialize the animation transformations for all nodes in a scene graph.
      * 
@@ -375,11 +382,11 @@ class Animation {
  * Set the Animation name
 -------------------------------------*/
 template <typename std_string_type>
-void Animation::set_anim_name(std_string_type&& name) noexcept {
+void Animation::set_anim_name(std_string_type&& name) noexcept
+{
     animationId = utils::string_hash(name.c_str());
     animName = std::forward<std_string_type>(name);
 }
-
 
 
 } // end draw namespace
